+++ src/api/server.php
@@ -117,6 +117,12 @@ switch ($QUERY[PQuery::ACTION]) {
   case PActions::SEND_HEALTH_CHECK:
     $api = new APISendHealthCheck();
     break;
+  /**
+   * Retrieve the found hashes of a hashlist
+   */
+  case PActions::GET_FOUND:
+    $api = new APIGetFound();
+    break;
 }
 
 DServerLog::log(DServerLog::TRACE, "Received from " . Util::getIP() . ": " . json_encode($QUERY));