+++ src/dba/models/Hashlist.class.php
@@ -15,8 +15,9 @@ class Hashlist extends AbstractModel {
   private $isSalted;
   private $accessGroupId;
   private $notes;
+  private $brainId;
   
-  function __construct($hashlistId, $hashlistName, $format, $hashTypeId, $hashCount, $saltSeparator, $cracked, $isSecret, $hexSalt, $isSalted, $accessGroupId, $notes) {
+  function __construct($hashlistId, $hashlistName, $format, $hashTypeId, $hashCount, $saltSeparator, $cracked, $isSecret, $hexSalt, $isSalted, $accessGroupId, $notes, $brainId) {
     $this->hashlistId = $hashlistId;
     $this->hashlistName = $hashlistName;
     $this->format = $format;
@@ -29,6 +30,7 @@ class Hashlist extends AbstractModel {
     $this->isSalted = $isSalted;
     $this->accessGroupId = $accessGroupId;
     $this->notes = $notes;
+    $this->brainId = $brainId;
   }
   
   function getKeyValueDict() {
@@ -45,6 +47,7 @@ class Hashlist extends AbstractModel {
     $dict['isSalted'] = $this->isSalted;
     $dict['accessGroupId'] = $this->accessGroupId;
     $dict['notes'] = $this->notes;
+    $dict['brainId'] = $this->brainId;
     
     return $dict;
   }
@@ -160,6 +163,14 @@ class Hashlist extends AbstractModel {
   function setNotes($notes){
     $this->notes = $notes;
   }
+  
+  function getBrainId(){
+    return $this->brainId;
+  }
+  
+  function setBrainId($brainId){
+    $this->brainId = $brainId;
+  }
 
   const HASHLIST_ID = "hashlistId";
   const HASHLIST_NAME = "hashlistName";
@@ -173,4 +184,5 @@ class Hashlist extends AbstractModel {
   const IS_SALTED = "isSalted";
   const ACCESS_GROUP_ID = "accessGroupId";
   const NOTES = "notes";
+  const BRAIN_ID = "brainId";
 }