+++ src/getFound.php
@@ -3,12 +3,10 @@
 use DBA\Agent;
 use DBA\ContainFilter;
 use DBA\Hash;
-use DBA\HashBinary;
 use DBA\Hashlist;
 use DBA\OrderFilter;
 use DBA\QueryFilter;
 use DBA\Factory;
-use DBA\Assignment;
 
 require_once(dirname(__FILE__) . "/inc/load.php");
 
@@ -42,13 +40,7 @@ if ($agent->getOs() == DOperatingSystem::WINDOWS) {
   $lineDelimiter = "\r\n";
 }
 
-$qF = new QueryFilter(Assignment::AGENT_ID, $agent->getId(), "=");
-$assignment = Factory::getAssignmentFactory()->filter([Factory::FILTER => $qF], true);
-$task = Factory::getTaskFactory()->get($assignment->getTaskId());
-
 $format = $hashlists[0]->getFormat();
-$brain = ($hashlists[0]->getBrainId() && !$task->getIsPrince() && !$task->getForcePipe()) ? true : false;
-$count = 0;
 switch ($format) {
   case DHashlistFormat::PLAIN:
     header_remove("Content-Type");
@@ -59,22 +51,16 @@ switch ($format) {
       do {
         $oF = new OrderFilter(Hash::HASH_ID, "ASC LIMIT $limit,$size");
         $qF1 = new QueryFilter(Hash::HASHLIST_ID, $hashlist->getId(), "=");
-        $qF2 = new QueryFilter(Hash::IS_CRACKED, 0, "=");
-        if($brain){
-          $current = Factory::getHashFactory()->filter([Factory::FILTER => $qF1, Factory::ORDER => $oF]);
-        }
-        else {
-          $current = Factory::getHashFactory()->filter([Factory::FILTER => [$qF1, $qF2], Factory::ORDER => $oF]);
-        }
+        $qF2 = new QueryFilter(Hash::IS_CRACKED, 1, "=");
+        $current = Factory::getHashFactory()->filter([Factory::FILTER => [$qF1, $qF2], Factory::ORDER => $oF]);
         
         $output = "";
-        $count += sizeof($current);
         foreach ($current as $entry) {
           $output .= $entry->getHash();
           if (strlen($entry->getSalt()) > 0) {
             $output .= "\t" . $entry->getSalt();
           }
-          $output .= $lineDelimiter;
+          $output .= "\tFF" . $lineDelimiter;
         }
         echo $output;
         
@@ -82,24 +68,4 @@ switch ($format) {
       } while (sizeof($current) > 0);
     }
     break;
-  case DHashlistFormat::BINARY:
-  case DHashlistFormat::WPA:
-    header_remove("Content-Type");
-    header('Content-Type: application/octet-stream');
-    $output = "";
-    foreach ($hashlists as $hashlist) {
-      $qF1 = new QueryFilter(HashBinary::HASHLIST_ID, $hashlist->getId(), "=");
-      $qF2 = new QueryFilter(HashBinary::IS_CRACKED, 0, "=");
-      $current = Factory::getHashBinaryFactory()->filter([Factory::FILTER => [$qF1, $qF2]]);
-      $count += sizeof($current);
-      foreach ($current as $entry) {
-        $output .= Util::hextobin($entry->getHash());
-      }
-    }
-    echo $output;
-    break;
-}
-
-if ($count == 0) {
-  die("No hashes are available to crack!");
 }